trinity-core/domain_integration.py
"""
MeeTARA Lab - Comprehensive Domain Integration
Aligns MCP Agent Ecosystem with all 62 domains and 10 Enhanced TARA Features
"""

import asyncio
import yaml
import json
from pathlib import Path
from typing import Dict, Any, List, Optional, Set
from datetime import datetime

class DomainIntegration:
    """Comprehensive integration of all 62 domains with 10 Enhanced TARA Features"""
    
    def __init__(self):
        # Load all 62 domains from cloud-optimized mapping
        self.domain_mapping = self._load_domain_mapping()
        
        # 10 Enhanced TARA Features - MUST BE PRESERVED
        self.enhanced_features = {
            "tts_manager": {
                "description": "6 voice categories with Edge-TTS + pyttsx3 integration",
                "file": "trinity-core/tts_manager.py",
                "domains": "ALL"
            },
            "emotion_detector": {
                "description": "RoBERTa-based emotion detection with professional context",
                "file": "trinity-core/emotion_detector.py", 
                "domains": "ALL"
            },
            "intelligent_router": {
                "description": "Multi-domain analysis with RoBERTa-powered routing",
                "file": "trinity-core/intelligent_router.py",
                "domains": "ALL"
            }
        }
        
    def _load_domain_mapping(self) -> Dict[str, Any]:
        """Load all 62 domains from cloud-optimized-domain-mapping.yaml"""
        try:
            config_path = Path("config/cloud-optimized-domain-mapping.yaml")
            with open(config_path, 'r') as f:
                mapping = yaml.safe_load(f)
                
            # Extract all domains
            all_domains = {}
            domain_categories = ["healthcare", "daily_life", "business", "education", "creative", "technology", "specialized"]
            
            for category in domain_categories:
                if category in mapping:
                    all_domains[category] = mapping[category]
                    
            return {
                "domains": all_domains,
                "total_domain_count": sum(len(domains) for domains in all_domains.values())
            }
            
        except Exception as e:
            print(f"⚠️ Could not load domain mapping: {e}")
            return {"domains": {}, "total_domain_count": 0}

# Global instance
domain_integration = DomainIntegration() 